changements

+++ cogs/errors.py
@@ -21,6 +21,8 @@ class ErrorCog(commands.Cog):
                 embed_erreur.set_footer(
                     text="ce message sera supprimé dans 15 secondes.")
                 msg = await ctx.reply(embed=embed_erreur)
+                if ctx.channel.type == discord.ChannelType.private:
+                    return
                 await asyncio.sleep(15)
                 await ctx.message.delete()
                 await msg.delete()
@@ -33,7 +35,7 @@ class ErrorCog(commands.Cog):
                     text="ce message sera supprimé dans 15 secondes.")
                 msg = await ctx.reply(embed=embed)
                 if ctx.channel.type == discord.ChannelType.private:
-                    await ctx.send("Message is in dm")
+                    return
                 await asyncio.sleep(15)
                 await ctx.message.delete()
                 await msg.delete()
